requirements.txt
pygithub=~1.55
requests
